renamed lfx-scheduler -> scheduler
