+++ scheduler/config/rbac/auth_proxy_client_clusterrole.yaml
