+++ scheduler/go.mod
@@ -1,4 +1,4 @@
-module github.com/keptn-sandbox/lifecycle-controller/lfc-scheduler
+module github.com/keptn-sandbox/lifecycle-controller/scheduler
 
 go 1.18
 
@@ -111,7 +111,7 @@ require (
 )
 
 replace (
-	github.com/keptn-sandbox/lifecycle-controller/lfc-scheduler/pkg/klcpermit => /pkg/klcpermit
+	github.com/keptn-sandbox/lifecycle-controller/scheduler/pkg/klcpermit => /pkg/klcpermit
 	k8s.io/api => k8s.io/api v0.24.3
 	k8s.io/apiextensions-apiserver => k8s.io/apiextensions-apiserver v0.24.3
 	k8s.io/apimachinery => k8s.io/apimachinery v0.24.3
