+++ scheduler/manifests/install/kustomization.yaml
@@ -1,0 +1,21 @@
+namespace: keptn-lifecycle-controller-system
+
+resources:
+  - base/deployment.yaml
+  - base/rbac.yaml
+  - base/serviceaccount.yaml
+
+generatorOptions:
+  disableNameSuffixHash: true
+
+configMapGenerator:
+  - files:
+      - base/scheduler-config.yaml
+    name: scheduler-config
+
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+images:
+  - name: klfc-scheduler
+    newName: ghcr.io/keptn-sandbox/scheduler
+    newTag: latest
